Assorted changes

+++ apps/3d_rec_framework/include/pcl/apps/3d_rec_framework/feature_wrapper/local/fpfh_local_estimator.h
@@ -75,7 +75,7 @@ public:
     fpfh_estimate.compute(*fpfhs);
 
     signatures->resize(fpfhs->size());
-    signatures->width = static_cast<int>(fpfhs->size());
+    signatures->width = fpfhs->size();
     signatures->height = 1;
 
     int size_feat = 33;
