+++ apps/3d_rec_framework/include/pcl/apps/3d_rec_framework/feature_wrapper/local/shot_local_estimator_omp.h
@@ -112,7 +112,7 @@ public:
     }
 
     signatures->resize(shots->size());
-    signatures->width = static_cast<int>(shots->size());
+    signatures->width = shots->size();
     signatures->height = 1;
 
     int size_feat = sizeof((*signatures)[0].histogram) / sizeof(float);
